Issue with mongo schema
