+++ server/.env.example
@@ -1,0 +1,2 @@
+PORT=yourport
+MONGO_URI=yourMongodbAtlasConnectionString
