+++ server/modules/user.controller.js
@@ -1,0 +1,27 @@
+import User from "./user.model.js";
+
+const getAllUser = () => {
+  return "getAllUser";
+};
+
+const getUser = () => {
+  return "getUser";
+};
+
+const loginUser = () => {
+  return "loginUser";
+};
+
+const registerUser = () => {
+  return "registerUser";
+};
+
+const deleteUser = () => {
+  return "deleteUser";
+};
+
+const updateUser = () => {
+  return "updateUser";
+};
+
+export { getUser, getAllUser, loginUser, registerUser, deleteUser, updateUser };
