+++ server/routes/apiRouter.js
@@ -1,0 +1,8 @@
+import { Router } from "express";
+import userRouter from "../modules/user.routes.js";
+
+const router = Router();
+
+router.use("/user", userRouter);
+
+export default router;
