+++ server/routes/index.js
@@ -1,5 +1,6 @@
 import { Router } from "express";
-import apiRouter from "./apiRouter";
+import apiRouter from "./apiRouter.js";
+
 const router = Router();
 
 router.use("/api/v1", apiRouter);
